requirements.txt
# Core dependencies
mcp>=1.20.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
yamllint>=1.35.0

# Testing
pytest>=7.4.0
